Refactor A2A and MCP Serving Structure
- Moved A2A and MCP related configurations and server implementations into their respective directories for better organization.
- Introduced new files for A2A agent card, executor, task manager, and envelope handling.
- Updated imports in existing files to reflect the new structure.
- Added initial configurations for MCP serving.

This change enhances the modularity and clarity of the serving components, making it easier to manage and extend in the future!
